requirements.txt
python-telegram-bot[job-queue]>=20.0
openai>=1.0
python-dotenv>=1.0
pyyaml>=6.0
trafilatura>=1.6
httpx>=0.24
pytz>=2023.3
lxml_html_clean
